requirements.txt
yt-dlp
mutagen
requests
Pillow
keyring
filelock
